backend/requirements.txt
fastapi
uvicorn
google-genai
python-pptx
pydantic
pydantic-settings
python-dotenv
langchain
langchain-community
langchain-google-genai
faiss-cpu
pypdf
unstructured
python-multipart
